requirements_lab1_postgres.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
ujson>=5.10.0
